tests/wpt/web-platform-tests/speculation-rules/prerender/restriction-local-file-system-access.https.html

<!DOCTYPE html>
<title>Same-origin prerendering cannot access local file system</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const uid = token();
  const bc = new PrerenderChannel('prerender-channel', uid);

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // Start prerendering a page that attempts to show a local file picker.
  startPrerendering(`resources/file-picker.html?uid=${uid}`);
  const result = await gotMessage;
  assert_equals(result, 'SecurityError')
}, `prerendering pages should not be able to access the local file system ` +
   `via the File System Access API`);

</script>
</body>
